requirements.txt
pandas>=2.0.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.0.0
openai==1.63.2
python-dotenv==1.0.7
peewee==3.17.9
Faker==36.1.1
typing-extensions==4.12.2
tqdm==4.66.4 